requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mediapipe-silicon==0.9.2.1
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
pillow==11.3.0
protobuf==3.20.3
pycparser==2.23
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scipy==1.13.1
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
torch==2.8.0
typing_extensions==4.15.0
zipp==3.23.0
